Request downloads

Request survey response downloads.

request_downloads(surveyIds)

download_requested(requests, format = "csv", saveDir = NULL,
  verbose = FALSE, ...)

# S3 method for qualtrics_download
download_requested(requests, format = "csv",
  saveDir = NULL, verbose = FALSE, ...)

Arguments

surveyIds
A vector of survey ids.

requests
Requests, as returned by request_downloads

format
file format json, by default (can be csv or tsv). We don't provide SPSS yet.

saveDir
path to a local directory to save the exported file. Default is a temporary file in rds format that is removed as the end of your session. Default name will be the surveyId.

verbose
prints progress bars for download (set to false by default)

...
Any other option.

Value

A request_downloads returns object of class qualtrics_download while download_requested returns a list.

Details

The Qualtrics API takes some time to prepare a downloads, hence download_requested taking a some time to download the requested data. What actually happens is that first a download is requested, after the request Qualtrics prepares a file that download_requested downloads when it is ready. It is therefore more efficient, when the user wants to download data from multiple surveys, more efficient to request each and every download before checking whether they are ready. Those performances will headvily depend on the number of surveys and the amount of data the ultimately user wants to obtain.

Functions

request_downloads: Request download (prepares files).

download_requested: Download requested files.

Examples

(requests <- request_downloads(c(1,2)))

#> Download Requests:
#>  ✔ 0  Successful
#>  ✖ 2  Unsuccessful

data <- download_requested(requests)

#> ✖ Not downloading unsuccessful surveyIds:
#>  1
#> 2