Create a new contact in a mailing list

create a new contact in a mailing list. In addition to the mailing list id, you can pass a named list of parameters to define your contact. All parameters are optional but empty parameters cannot be passed. Parameters are: firstName, lastName, email, externalDataRef, language, unsubscribed and embeddedData. The embeddedData element must a string representation of a JSON object.

create_contact(mailinglist_id, contact_options)

Arguments

mailinglist_id
ID of the mailing list in which to create the contact

contact_options
list of new contact information

Value

The id of the created mailing list

Examples

if (FALSE) {
new_contact <- list(
"firstName" = "James",
"lastName" = "Paddignton",
"email" = "james.pad@gmailing.com")
new_contact_id <- create_contact("ML_0HT4q2Ni634kLhH", new_contact)}