Create a reminder distribution for an existing email distribution

Based on an existing email distribution, create a reminder to take a survey for recipients with non completed surveys. Recipients who have opted out will not receive the reminder.

create_reminder_distribution(
  parent_distribution_id,
  library_id,
  message_id,
  subject = "Reminder - Participate to the survey",
  send_date = paste0(Sys.Date() + 2, "T00:00:00Z"),
  from_email = "noreply@qualtrics.com",
  from_name = "Qualtrics",
  reply_to_email = "noreply@qualtrics.com"
)

Arguments

parent_distribution_id
id of the library item

library_id
Library ID of the message

message_id
id of the message item

subject
email subject, default is "Reminder - Participate to the survey"

send_date
date for distribution to be sent (default to time + 1 day)

from_email
sender email, default is noreply@qualtrics.com

from_name
appearing sender name, default is Qualtrics

reply_to_email
reply email, default is noreply@qualtrics.com

Value

The reminder distribution id

Details

The parent distribution id can be found using the 'list_distributions' call. The sender email parameter can only be sent from another email if the organization authorizes a send on its behalf.

Examples

if (FALSE) {
  create_reminder_distribution(
    "EMD_6yYmF6vopfUz4gn",
    library_id = "UR_0NXtl92JJWqfWcJ",
    message_id = "MS_0fddN2xI3J0nGQZ"
)
}