Create a new survey distribution via SMS

Create a new survey distribution via SMS

create_sms_distribution(
  send_date,
  method,
  survey_id,
  name,
  mailinglist_id,
  message_id = NULL,
  library_id = NULL,
  messagetext = NULL,
  contact_id = NULL,
  parentdistribution_id = NULL
)

Arguments

send_date
date for distribution to be sent

method
Determines how the survey will be sent out via SMS. Can be either Invite, Interactive, Reminder or Thankyou.

survey_id
the id of survey to copy

name
Name for the SMS distribution

mailinglist_id
Mailing List ID for a batch distribution - Required for Invite and Interactive method. Cannot be provided for Reminder and Thankyou

message_id
ID of the message. The messageId is required in conjunction with the libraryId if no messageText is supplied. A message object is only required with the Invite, Reminder, and Thankyou methods.

library_id
Library ID of the message. libraryId is required in conjunction with the messageId if no messageText is supplied. A message object is only required by "Invite" method.

messagetext
Custom message text. Either messageText or messageId must be provided but not both. A message object is only required with the Invite, Reminder, and Thankyou methods.

contact_id
Contact ID for an individual distribution. Cannot be provided for Reminder and Thankyou

parentdistribution_id
ID of the distribution to send a Reminder or Thankyou distribution. Can be used only with these two methods.

Value

A list