Export a survey's responses into R

Export survey responses into R and/or save exported file. Currently single exports calls are limited to 1.8 GB. Above that size, we recommend you split your export accross different date ranges

get_survey_responses_new(survey_id, format = "json", verbose = FALSE,
  saveDir = NULL, filename = NULL, ...)

Arguments

survey_id
the id of survey to copy

format
file format json, by default (can be csv or tsv). We don't provide SPSS yet.

verbose
default FALSE

saveDir
path to a local directory to save the exported file. Default is a temporary file in rds format that is removed as the end of your session. Default name will be the surveyId.

filename
specify filename for saving. If NULL, uses the survey id

...
a list of named parameters, see https://api.qualtrics.com/docs/getting-survey-responses-new for parameter names

Value

A status code

Examples

# NOT RUN {
get_survey_responses("SV_012345678901234", "csv")
# }# NOT RUN {
get_survey_responses("SV_012345678901234", "csv", filename = "name_export")
# }# NOT RUN {
mysurvey <- getSurvey(surveyID = "surveyId",
                      startDate = "2017-01-01",
                      endDate = "2017-01-31",
                      limit = 100,
                      useLabels = TRUE,
                      seenUnansweredRecode = "UNANS",
                      verbose = TRUE)

# }