Send Email to Mailing List

You can use this call to send an email to a mailing list without having to create the email message in Qualtrics first.

send_email_mailinglist(
  mailingListId,
  sendDate = paste0(Sys.Date() + 1, "T00:00:00Z"),
  subject = "Take this survey",
  messageText = "Hello, take this survey",
  fromEmail = "noreply@qualtrics.com",
  fromName = "Qualtrics",
  replyToEmail = "noreply@qualtrics.com"
)

Arguments

mailingListId
the id of survey

sendDate
the date to send the email (default is +1 day)

subject
the email subject

messageText
email text body

fromEmail
sender email, default is noreply@qualtrics.com

fromName
the send name default is Qualtrics

replyToEmail
reply email, default is noreply@qualtrics.com

Value

The email distribution id

Examples

if (FALSE) {
 create_distribution(
   "ML_7aoriSKinHh8MfP",
   "john.doe@qualtrics.com",
   "John Doe",
   "Participate in this awesome survey")
}