Request downloads

Request survey response downloads.

request_downloads(
  surveyIds,
  format = "json",
  ...,
  verbose = interactive(),
  retryOnRateLimit = TRUE
)

download_requested(
  requests,
  saveDir = NULL,
  verbose = interactive(),
  retryOnRateLimit = TRUE
)

# S3 method for qualtrics_download
download_requested(
  requests,
  saveDir = NULL,
  verbose = interactive(),
  retryOnRateLimit = TRUE
)

check_status(requests)

Arguments

surveyIds
A vector of survey ids.

format
file format json, by default (can be csv or tsv). We don't provide SPSS yet.

...
a vector of named parameters, see https://api.qualtrics.com/reference Create Response Export for parameter names.

verbose
print progress for heavy downloads (default is FALSE)

retryOnRateLimit
Whether to retry if initials API call fails.

requests
the list output of request_downloads

saveDir
the output dir to save the data

Value

A request_downloads returns object of class qualtrics_download while download_requested returns a list.

Details

The Qualtrics API downloads survey responses in several steps. First a download is requested. Qualtrics prepares the file and download_requested can be triggered to download the file when it is ready. When downloading responses from a large nuimber of surveys, it is more efficient to request all downloads if a first place, check that the downloads are ready and then retrieve all the files. The performances will headvily depend on the number of surveys and the amount of data the user ultimately wants to obtain.

Functions

request_downloads: Request download (prepares files).

download_requested: Download requested files.

Examples