Export a survey's responses into R

Export survey responses into R and/or save exported file. Currently single exports calls are limited to 1.8 GB. Above that size, we recommend you split your export accross different date ranges. When exporting a large number of surveys successively, we recommend using the more efficient request_downloads and download_requested functions.

get_survey_responses(
  survey_id,
  format = "csv",
  verbose = FALSE,
  saveDir = NULL,
  filename = NULL,
  ...
)

Arguments

survey_id
the id of survey to copy

format
file format csv, by default (can be json or tsv). We don't provide SPSS yet.

verbose
default FALSE

saveDir
path to a local directory to save the exported file. Default is a temporary file in rds format that is removed at the end of your session. Default name will be the surveyId.

filename
specify filename for saving. If NULL, uses the survey id

...
a vector of named parameters, see all parameters in reference documentation.

Value

A status code

Examples

if (FALSE) get_survey_responses("SV_012345678901234", "csv")
if (FALSE) get_survey_responses("SV_012345678901234", format = "csv",verbose = TRUE,
saveDir = "./", filename = "name_export", useLabels = TRUE, limit = 10)